Fix getPlayerChoice function for bad input
After a round with a bad user input, the following round would
not play correctly. Variable playerSelection would return as
undefined again even with a correct input and the round would not
count towards the score.

Function game() displays a message to announce the winner of the best
of 5 games.